Enrich api proposals (#278)
* add dispute votes to api

* fmt

* return milestones in voting round

* fmt

* clippy

* fix benchmark build

* fmt

+++ pallets/proposals/src/tests/refunds.rs
@@ -273,7 +273,7 @@ fn refund_check_refund_amount() {
         let charlie_pre_creation =
             <Test as Config>::MultiCurrency::free_balance(CurrencyId::Native, &CHARLIE);
         let per_contribution = 100000u128;
-        let contributions = get_contributions::<Test>(vec![BOB, CHARLIE], per_contribution as u128);
+        let contributions = get_contributions::<Test>(vec![BOB, CHARLIE], per_contribution);
         let milestones = get_milestones(10);
         let project_key = create_and_fund_project::<Test>(
             ALICE,
@@ -331,7 +331,7 @@ fn refund_takes_imbue_fee() {
         );
         let per_contribution = 500000u128;
 
-        let contributions = get_contributions::<Test>(vec![BOB, CHARLIE], per_contribution as u128);
+        let contributions = get_contributions::<Test>(vec![BOB, CHARLIE], per_contribution);
         let milestones = get_milestones(10);
         let project_key = create_and_fund_project::<Test>(
             ALICE,
